Add enforcement module.
This is not fully fleshed-out. I'm in the middle of creating it
and realizing that my data model is not quite right. I just want
to commit what I have so that I can fix the data model before
I try to move this forward.

For now, the enforcement module is the beginnings of the logic
that will determine what the kid has done with their sessions
and take the appropriate action to inform them that their
session will be forcibly ended and that they are out of time.