Add logic for getting a single Model row.
I also made it so that I'm specifying the ID column differently,
in a way that inspecting the definition is easier. I was going
to use this to validate that the user can not-specify the id field,
but when it came down to it I found it easier to just not do
any validation at all and assume the caller knows what they
need to specify. SQLite will complain if they don't.