Get a proof-of-concept client daemon working.
This daemon requests the list of important programs, takes a snapshot
of its own running programs, uploads that snapshot, and gets a list
of actions based on the snapshot and performs them.

That's everything I need it to do.

The server will need a lot of work, but for now, this is installable.